# Conflicts:
#	TASK.md
#	src/main/java/ua/duikt/learning/java/pro/spring/controllers/AttachmentController.java
#	src/main/java/ua/duikt/learning/java/pro/spring/controllers/CommentController.java
#	src/main/java/ua/duikt/learning/java/pro/spring/controllers/LabelController.java
#	src/main/java/ua/duikt/learning/java/pro/spring/controllers/StatusController.java
#	src/main/java/ua/duikt/learning/java/pro/spring/entity/Attachment.java
#	src/main/java/ua/duikt/learning/java/pro/spring/entity/Issue.java
#	src/main/java/ua/duikt/learning/java/pro/spring/entity/IssueComment.java
#	src/main/java/ua/duikt/learning/java/pro/spring/entity/IssueHistory.java
#	src/main/java/ua/duikt/learning/java/pro/spring/entity/IssueLabel.java
#	src/main/java/ua/duikt/learning/java/pro/spring/entity/Label.java
#	src/main/java/ua/duikt/learning/java/pro/spring/entity/Project.java
#	src/main/java/ua/duikt/learning/java/pro/spring/entity/ProjectMember.java
#	src/main/java/ua/duikt/learning/java/pro/spring/entity/Role.java
#	src/main/java/ua/duikt/learning/java/pro/spring/entity/Status.java
#	src/main/java/ua/duikt/learning/java/pro/spring/entity/User.java
#	src/main/java/ua/duikt/learning/java/pro/spring/entity/UserRole.java
#	src/main/java/ua/duikt/learning/java/pro/spring/service/IssueService.java
#	src/main/java/ua/duikt/learning/java/pro/spring/service/impl/DetailsServiceImpl.java
#	src/main/java/ua/duikt/learning/java/pro/spring/service/impl/IssueServiceImpl.java
#	src/main/java/ua/duikt/learning/java/pro/spring/service/impl/UserServiceImpl.java
#	src/test/java/ua/duikt/learning/java/pro/spring/sprint02/service/DetailsServiceTest.java
#	src/test/java/ua/duikt/learning/java/pro/spring/sprint03/controllers/AttachmentControllerTest.java
#	src/test/java/ua/duikt/learning/java/pro/spring/sprint03/controllers/CommentControllerTest.java
#	src/test/java/ua/duikt/learning/java/pro/spring/sprint03/controllers/LabelControllerTest.java
#	src/test/java/ua/duikt/learning/java/pro/spring/sprint03/controllers/StatusControllerTest.java
#	src/test/java/ua/duikt/learning/java/pro/spring/sprint03/service/IssueServiceTest.java

+++ src/main/java/ua/duikt/learning/java/pro/spring/controllers/AttachmentController.java
@@ -22,7 +22,7 @@ public class AttachmentController {
     private final AttachmentService attachmentService;
 
     @PostMapping("/issues/{issueId}/attachments")
-    public ResponseEntity<String> addAttachment(@PathVariable Integer issueId,
+    public ResponseEntity<String> addAttachment(@PathVariable Long issueId,
                                                 @RequestBody AddAttachmentRequest request) {
         boolean added = attachmentService.addAttachment(
                 issueId,
@@ -38,12 +38,12 @@ public class AttachmentController {
     }
 
     @GetMapping("/issues/{issueId}/attachments")
-    public ResponseEntity<List<Attachment>> getAttachments(@PathVariable Integer issueId) {
+    public ResponseEntity<List<Attachment>> getAttachments(@PathVariable Long issueId) {
         return ResponseEntity.ok(attachmentService.getAttachments(issueId));
     }
 
     @DeleteMapping("/attachments/{attachmentId}")
-    public ResponseEntity<Void> deleteAttachment(@PathVariable Integer attachmentId) {
+    public ResponseEntity<Void> deleteAttachment(@PathVariable Long attachmentId) {
         boolean deleted = attachmentService.deleteAttachment(attachmentId);
         return deleted ? ResponseEntity.noContent().build() : ResponseEntity.notFound().build();
     }
