+++ src/main/java/ua/duikt/learning/java/pro/spring/service/impl/RoleServiceImpl.java
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * Created by Mykyta Sirobaba on 13.01.2026.
@@ -17,12 +17,12 @@ import java.util.concurrent.atomic.AtomicInteger;
 @Service
 public class RoleServiceImpl implements RoleService {
 
-    private final Map<Integer, Role> roleTable = new ConcurrentHashMap<>();
-    private final AtomicInteger idGenerator = new AtomicInteger(1);
+    private final Map<Long, Role> roleTable = new ConcurrentHashMap<>();
+    private final AtomicLong idGenerator = new AtomicLong(1);
 
     @Override
-    public Integer createRole(String name) {
-        Integer id = idGenerator.getAndIncrement();
+    public Long createRole(String name) {
+        Long id = idGenerator.getAndIncrement();
         Role role = Role.builder().id(id).name(name).build();
         roleTable.put(id, role);
         return id;
